Formatting, mumbai and ABIs copy-paste script.

+++ contracts/NFTFactory.sol
@@ -1,18 +1,13 @@
 // SPDX-License-Identifier: MIT
-
 pragma solidity ^0.8.0;
-
-import "hardhat/console.sol";
 
 contract NFTFactory {
     // Structure for storing nft properties
     struct nftProps {
         string nftKey; // Key is unique string and concatenation of nftAddress and nftId
-
         address nftOwner;
         address nftAddress;
         uint256 nftId;
-
         string nftName;
         string nftImageURL;
     }
@@ -21,13 +16,13 @@ contract NFTFactory {
     address[] internal userAddressList;
 
     // Mapping of above nft key to nftProps
-    mapping (string => nftProps) public nftKeyToNftProps;
+    mapping(string => nftProps) public nftKeyToNftProps;
 
-    function getContractBalance() external view returns(uint) {
+    function getContractBalance() external view returns (uint256) {
         return address(this).balance;
     }
 
-    function getUserAddressList() public view returns(address[] memory) {
+    function getUserAddressList() public view returns (address[] memory) {
         return userAddressList;
     }
 }
